code review (part 2)
